Quest5InProgress
Added semantic + 1 day for dates. When we have cur.d. and pay,d === (15 = 15) in return we have "1day left" now. And when we have cur = 14 and pay = 15 we divide 1/2 and haves 1(2) days left to payday. Good calculating option added.
Now needs add that into beforeDay() and other places where that's can needs be.